requirements.txt
boto3
fastapi
uvicorn
Pillow